+++ R/labels.R
@@ -65,7 +65,7 @@ assign_labels.data.frame <- function(x, ...) {
   ma <- match(nm, colnames(x), nomatch = NA_integer_)
 
   if (anyNA(ma)) {
-    stop("Columns not found: ", collapse(nm[is.na(ma)], sep = ", "), call. = FALSE)
+    stop("Columns not found: ", collapse0(nm[is.na(ma)], sep = ", "), call. = FALSE)
   }
 
   for (i in seq_along(nm)) {
@@ -141,7 +141,7 @@ remove_labels.data.frame <- function(x, cols, ...) {
 
     if (any(bad)) {
       stop("Column not found in data.frame:\n  ",
-           collapse(cols[bad], sep = ", "),
+           collapse0(cols[bad], sep = ", "),
            call. = FALSE)
     }
   }