+++ R/names.R
@@ -22,13 +22,20 @@ sort_names <- function(x) {
 #' @param x A vector of values
 #' @param nm A vector of names
 #' @export
-set_names <- function(x, nm = x) {
+set_names0 <- function(x, nm = x) {
   names(x) <- nm
   x
 }
 
-#' @rdname set_names
+#' @rdname set_names0
+#' @export
+set_names <- function(x, nm = x) {
+  .Deprecated("set_names0")
+  set_names0(x, nm)
+}
+
+#' @rdname set_names0
 #' @export
 remove_names <- function(x) {
-  set_names(x, NULL)
+  set_names0(x, NULL)
 }