+++ tests/testthat/test-str-utils.R
@@ -4,13 +4,13 @@ test_that("string slices", {
 
   # By length
   res <- str_slice(x, n = len)
-  expect_equal(paste(res, collapse = ""), x, info = "collapsed res")
+  expect_equal(collapse0(res, sep = ""), x, info = "collapsed res")
   lengths <- vapply(res, nchar, integer(1))
   expect_true(all(lengths <= len), info = "All lines within desired length")
 
   # By length and word
   res <- str_slice_by_word(x, n = len)
-  expect_equal(paste(res, collapse = ""), x, info = "By word:  Collapsed res")
+  expect_equal(collapse0(res, sep = ""), x, info = "By word:  Collapsed res")
   lengths <- vapply(res, nchar, integer(1))
   expect_true(all(lengths <= len), info = "By word:  All lines within desired length")